+++ DebugFileTools/SymbolsLib/symbol_file_static_temp.h
@@ -2,9 +2,9 @@
 
 struct s_symbol_file_static_temp
 {
-	uint32_t section_index;
-	uint32_t rva;
-	uint64_t rva_plus_base;
+	unsigned long section_index;
+	unsigned long rva;
+	unsigned long long rva_plus_base;
 	char lib_and_object[256];
 
 	char symbol_name[1024];
