+++ Tags/TagCodegen/high_level_tag_source_generator.h
@@ -9,7 +9,7 @@ public:
 	const blofeld::s_tag_group* get_tag_struct_tag_group(const blofeld::s_tag_struct_definition& tag_struct_definition) const;
 	void generate_header() const;
 	void generate_tag_constructor_params(std::stringstream& stream, const blofeld::s_tag_struct_definition& tag_struct_definition) const;
-	void generate_ctor_source(uint32_t source_index, uint32_t source_count) const;
+	void generate_ctor_source(unsigned long source_index, unsigned long source_count) const;
 	void generate_source_virtual() const;
 	void generate_source_misc() const;
