+++ Tags/TagDefinitions/types/field_type.h
@@ -104,8 +104,8 @@ namespace blofeld
 		k_number_of_blofeld_field_types
 	};
 
-	uint32_t get_blofeld_field_size(e_platform_type platform_type, e_field field);
-	uint32_t get_blofeld_field_size(const s_tag_field& field, s_engine_platform_build engine_platform_build);
+	unsigned long get_blofeld_field_size(e_platform_type platform_type, e_field field);
+	unsigned long get_blofeld_field_size(const s_tag_field& field, s_engine_platform_build engine_platform_build);
 	const char* field_to_string(e_field field);
 
 	enum e_field_id
