+++ Tags/TagDefinitions/types/tag_array_definition.h
@@ -7,7 +7,7 @@ namespace blofeld
 		const char* const name;
 		const char* const display_name;
 		const char* const filename;
-		int32_t const line;
+		long const line;
 		//unsigned long const count;
 		t_max_count_func* count;
 		const char* const count_string;
