+++ Tags/TagDefinitions/types/tag_block_definition.cpp
@@ -4,7 +4,7 @@ blofeld::s_tag_block_definition::s_tag_block_definition(
 	const char* const name,
 	const char* const display_name,
 	const char* const filename,
-	int32_t const line,
+	long const line,
 	t_max_count_func* max_count,
 	const char* const max_count_string,
 	const s_tag_struct_definition& struct_definition
