+++ Tags/TagDefinitions/types/tag_struct_definition.cpp
@@ -7,7 +7,7 @@ s_tag_struct_definition::s_tag_struct_definition(
 	const char* name,
 	const char* struct_name,
 	const char* filename,
-	int32_t const line,
+	long const line,
 	s_tag_persistent_identifier persistent_identifier,
 	s_tag_field* fields,
 	int alignment_bits
@@ -33,7 +33,7 @@ s_tag_struct_definition::s_tag_struct_definition(
 	const char* name,
 	const char* struct_name,
 	const char* filename,
-	int32_t const line,
+	long const line,
 	c_flags<e_tag_field_set_bit> runtime_flags,
 	c_tag_memory_attributes memory_attributes,
 	s_tag_persistent_identifier persistent_identifier,
@@ -44,7 +44,7 @@ s_tag_struct_definition::s_tag_struct_definition(
 	//const char(&name)[],
 	//const char(&struct_name)[],
 	//const char(&filename)[],
-	//int32_t const line,
+	//long const line,
 	//s_tag_persistent_identifier persistent_identifier,
 	//s_tag_field(&fields)[],
 	//int alignment_bits = 0
